Fix spec compliance: uint64 counters in /status.json, security in wifi_status
Spec 30 requires frames_in/dropped/sent and uptime_ms as uint64 in
/status.json. All counters and the frames_captured tracker were uint32_t,
causing technically non-compliant JSON output. Changed to uint64_t with
PRIu64 format specifiers throughout http_server, uvc_capture, and metrics.

Spec 20 requires the wifi_status CLI to show security mode. Added
"Security: WPA2-PSK" line to the command output.

+++ examples/usb_cam_stream/http_server.h
@@ -6,9 +6,9 @@
 
 /* Streaming counters (read by /status.json and metrics) */
 typedef struct {
-    volatile uint32_t frames_dequeued;
-    volatile uint32_t frames_dropped;
-    volatile uint32_t frames_sent;
+    volatile uint64_t frames_dequeued;
+    volatile uint64_t frames_dropped;
+    volatile uint64_t frames_sent;
     volatile bool     stream_client_connected;
 } http_counters_t;
